Script to generate comparison image for LEA/STRUCTURE
This script, which isn't in the README, isn't meant to be used generically like `LEA.R` and `STRUCTURE.R`. However, I've included documentation in the script that indicates what lines users would need to change, verify, or make decisions about should someone else need to use it for another comparison.